requirements.txt
flask
requests
websocket-client
numpy
krakenex
psycopg2-binary
openai